+++ src/lib.rs
@@ -52,5 +52,7 @@ pub use lru_cache::LruCache;
 pub use color_space::*;
 pub use color_ext::ToRgbComponents;
 
+pub(crate) use effect::ShaderExt;
+
 #[cfg(all(feature = "std-duration", feature = "web-time"))]
 compile_error!("Features 'std-duration' and 'web-time' cannot be enabled simultaneously");